Title: 1까지 나누기, Time: 136ms, Memory: 10MB, Status: Passed - CodeTree

+++ 231114/1까지 나누기/divide-by-1.java
@@ -5,7 +5,7 @@ public class Main {
         Scanner sc = new Scanner(System.in);
         int n = sc.nextInt();
         int i = 1;
-        double dValue = n;
+        int dValue = n;
         while(true){
             dValue /= i;
             if(dValue <= 1){
